heap and neighbor precalc
min heap implemented to allow for quickly finding the lowest fcost node. This turns a linear calculation in a logorithmic one. The neighbors of nodes are now precalculated, which cuts a lot of duplication of work. Based on 1000 executions, the heap improved the speed by 50% and the neighbor precalc by 15%